test: use tags to separate windows and unix tests to remove failing windows tests
